requirements.txt
# Core Framework Dependencies
streamlit>=1.28.0          # Web UI framework
langchain>=0.0.310         # Language model framework
langgraph>=0.0.30          # Workflow orchestration
openai>=1.3.0              # OpenAI API integration
pydantic>=2.5.0            # Data validation

# Agent API Integrations
requests>=2.31.0           # HTTP requests
python-dotenv>=1.0.0       # Environment variable management
serpapi>=0.1.0             # SerpAPI for Google search 
google-search-results>=2.4.2  # SERP API wrapper (alternative)
wikipedia>=1.4.0           # Wikipedia API access
euriai>=0.1.0              # EURI API for text summarization

# Graph & Visualization
networkx>=3.2.1            # Graph operations
plotly>=5.18.0             # Interactive visualizations
matplotlib>=3.8.0          # Static visualizations
graphviz>=0.20.1           # Graph generation

# Data Processing
pandas>=2.1.1              # Data manipulation
numpy>=1.26.0              # Numerical operations

# Image Processing
Pillow>=10.1.0             # Image handling

# Utilities
typing-extensions>=4.8.0   # Enhanced type hints
uuid>=1.30                 # Unique identifiers
beautifulsoup4>=4.12.2     # HTML parsing
markupsafe>=2.1.3          # Safe HTML

# LLM Integration
transformers>=4.34.0       # Hugging Face transformers (optional)
torch>=2.1.0               # PyTorch for LLM inference (optional)

# Optional: Performance Optimization
ujson>=5.8.0               # Fast JSON processing
aiohttp>=3.8.6             # Async HTTP